Refactored whole project [userId removed ]

+++ src/main/kotlin/example/com/routes/CommentRoutes.kt
@@ -17,8 +17,7 @@ import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 fun Route.createComment(
-    commentService: CommentService,
-    userService: UserService
+    commentService: CommentService
 ) {
     authenticate {
         post("/api/comment/create") {
@@ -27,39 +26,34 @@ fun Route.createComment(
                 return@post
             }
 
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                when (commentService.createComment(request)){
-                    CommentService.ValidationEvents.ErrorCommentTooLong -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.COMMENT_TOO_LONG
+            when (commentService.createComment(request,call.userId)){
+                CommentService.ValidationEvents.ErrorCommentTooLong -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.COMMENT_TOO_LONG
 
-                            )
                         )
-                    }
-                    CommentService.ValidationEvents.FieldEmpty -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.FIELDS_BLANK
+                    )
+                }
+                CommentService.ValidationEvents.FieldEmpty -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.FIELDS_BLANK
 
-                            )
                         )
-                    }
-                    CommentService.ValidationEvents.Success -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = true,
-                            )
+                    )
+                }
+                CommentService.ValidationEvents.Success -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = true,
                         )
-                    }
+                    )
                 }
             }
         }
@@ -87,7 +81,6 @@ fun Route.getCommentsForPost(
 
 fun Route.deleteComment(
     commentService: CommentService,
-    userService: UserService,
     likeService:LikeService
 ) {
     authenticate {
@@ -96,28 +89,27 @@ fun Route.deleteComment(
                 call.respond(HttpStatusCode.BadRequest)
                 return@delete
             }
-
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                 val deleted = commentService.deleteComment(request.commentId)
-                if (deleted) {
-                    likeService.deleteLikesForParent(request.commentId)
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = true
-                        )
+            val comment = commentService.getCommentById(request.commentId)
+            if(comment?.userId != call.userId) {
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
+            }
+            val deleted = commentService.deleteComment(request.commentId)
+            if (deleted) {
+                likeService.deleteLikesForParent(request.commentId)
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = true
                     )
-                }else {
-                    call.respond(
-                        HttpStatusCode.NotFound,
-                        BasicApiResponse(
-                            successful = false
-                        )
+                )
+            }else {
+                call.respond(
+                    HttpStatusCode.NotFound,
+                    BasicApiResponse(
+                        successful = false
                     )
-                }
+                )
             }
         }
     }
